backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
pydantic[email]
python-jose[cryptography]
passlib
bcrypt
pandas
openpyxl
python-multipart
pydantic-settings
argon2-cffi
